ref: move github_enterprise testutils to fixtures
